refactor: styled-components 분리
