+++ src/Pages/Main/Slider/StoreEditorSlider.tsx
@@ -1,58 +1,19 @@
-import styled from "styled-components";
-import "slick-carousel/slick/slick.css";
-import "slick-carousel/slick/slick-theme.css";
-import { randomImg } from "../MainImg";
 import {
-  ColumnBox,
-  RactTitle,
-  RandomRactangle,
-  SubRactangle,
-} from "../../Search/SearchStyle";
-import { StorySlider } from "./StoreStorySlider";
-import { useRecoilValue, useSetRecoilState } from "recoil";
-import { EditorDataState } from "../../../Atom/Main";
+  EditorBox,
+  EditorContainer,
+  EditorSlider,
+  EditorStoreTitle,
+  EditorTitle,
+  SubEditorBox,
+} from "./StyledSlider";
+import { randomImg } from "../MainImg";
+import { ColumnBox } from "../../Search/SearchStyle";
 import { useEffect, useState } from "react";
-import { isValidImage } from "./StoreListSlider";
-import { selectedRestaurantState } from "../../../Atom/Search";
 import { useNavigate } from "react-router-dom";
-
-const Container = styled.div`
-  display: flex;
-  flex-direction: column;
-  margin: 0 auto;
-`;
-
-const Title = styled.span`
-  margin-top: 10vh;
-  font-size: 25px;
-  font-weight: 600;
-`;
-
-const StoreTitle = styled(RactTitle)`
-  display: flex;
-  justify-content: center;
-`;
-
-const EditorBox = styled(RandomRactangle)`
-  display: flex;
-  width: ${(props) => props.width}px;
-  height: ${(props) => props.height}px;
-  position: relative;
-  border-radius: 20px;
-  background-size: cover;
-`;
-
-const SubEditorBox = styled(SubRactangle)`
-  position: absolute;
-  top: 80%;
-  left: 50%;
-  transform: translate(-50%, -50%);
-  height: ${(props) => props.height}px;
-  text-align: center;
-  justify-content: center;
-`;
-
-export const EditorSlider = styled(StorySlider)``;
+import { EditorDataState } from "../../../Atom/Main";
+import { selectedRestaurantState } from "../../../Atom/Search";
+import { useRecoilValue, useSetRecoilState } from "recoil";
+import { isValidImage } from "./StoreListSlider";
 
 export const StoreEditorSlider = () => {
   const editorData = useRecoilValue(EditorDataState);
@@ -71,19 +32,21 @@ export const StoreEditorSlider = () => {
       setImageUrls(urls); // 검사된 URL들을 상태에 저장
     });
   }, [editorData]);
+
   const settings = {
     dots: true,
     infinite: false,
     speed: 1000,
     slidesToShow: 4,
     slidesToScroll: 1,
+    arrows: true,
   };
 
   return (
     <>
-      <Container>
-        <Title>에디터가 선정한 역곡 식당</Title>
-      </Container>
+      <EditorContainer>
+        <EditorTitle>에디터가 선정한 역곡 식당</EditorTitle>
+      </EditorContainer>
       <ColumnBox>
         <EditorSlider {...settings}>
           {editorData.map((item, index) => (
@@ -98,7 +61,7 @@ export const StoreEditorSlider = () => {
               }}
             >
               <SubEditorBox width={0} height={50} imageURL="">
-                <StoreTitle>{item.placeName}</StoreTitle>
+                <EditorStoreTitle>{item.placeName}</EditorStoreTitle>
               </SubEditorBox>
             </EditorBox>
           ))}
